NEC-29 "Support publishing via NHR"
Added a new config option - SMP Integration. By default it is set to "Do not publish to SMP". It can be set to "Publish to configured SMP" only if AS2 or AS4 access point is defined and at least one profile: "Setup is not ready for publishing: either AS2 or AS4 access point should be selected and at least one profile. SMP integration is switched back to 'Do not publish to SMP'."

Added validation, that it is not possible to save setup if above condition is not met and "Publish to configured SMP " is selected.

Changed the way identifiers statistics is shown per organisation - for not published, it contains only number of active identifiers.

When new identifiers are loaded in mode "Do not publish" - they do not get PENDING publishing state.

Added warning 'Are you sure, that you want to disable publishing to SMP? All active and published identifiers will be deleted from SMP.'.

Forbidden to change SMP-related data, if any identifier is in PENDING state: "Setup is not ready for publishing: either AS2 or AS4 access point should be selected and at least one profile. SMP integration is switched back to 'Do not publish to SMP'."

When SMP configuration is changed, also FAILED previously identifiers are scheduled for publishing.